Initial Development Commit
I don't really know where to start with this but here goes, there is a
MAJOR amount of work still required to bring RawPhone even close to what is
hoped for but I think this gives a solid base to start with even if it is still
very rough! :)

This commit makes a number of significant changes to the original base
RawPhone application each of which will be explained in detail below IF
I can remember them all...

1. Device information was split into a new class to allow easy modifications
in the future, all items such as IMEI, Operator etc are now contained within
this class.

2. RootTools library added to provide access to helper functions such as
checking for the provision of root, and offering installation of BusyBox if
it is not detected.

3. CMDProcessor - The AOKP CMDProcessor has been added to RawPhone allowing
shell commands and various helper functions to be executed, I find this library
to be very stable and it works very well. Some items already added with this
include checking for Busybox installation and the util Microcom which should
allow some form of serial communication on the device (NOT TESTED YET!).

4. Microcom applet - This is an applet available through Busybox which has been
included and RawPhone will prompt to install this if it is not located, initial
reading regarding this points to the fact it is supposed to allow for the
issuing of basic AT serial commands ON THE DEVICE. Very interesting indeed but
as yet this is untested as I have not written the methods to issue the AT
commands using the applet.

5. Initial changes have also been made for a revamp of the UI but as yet this
has not been implemented.