OemCommands Updated
Included partial OemCommands class as found in the SamsungServiceMode application
which has allowed for activating the Ciphering Indicator RIL response, although
this is specific to Samsung it does at least pave the way for further commands
and this interface uses the invokeOemRilRequestRaw method.

Use of this method has required the reapplication of the frameworks libs to
allow access to the hidden APIs contained within the SDK, the relfection methods
have not been removed as they will still probably be very useful.

Extensive changes have also been made to the manifest to allow the application
to share the com.android.phone process, this also requires the signing of the
application with the SDK platform key.

As signing is required from this point forward I will attach a new github
release apk to match the commits pushed, this will hopefully allow for easy
testing as people will not need at import/sign with the platform keys.