Reduce target sdk version & code updates
Minimum Android SDK version reduced to SDK 9 - Gingerbread to target a wider range of devices,
a number of code changes were required to comply with the SDK 9 level requirements.

Further changes made to the ATRilHook class in an attempt to access the invokeOemRilRequestStrings method
correctly, also some updates made to the handling of OemRilRaw requests and their response. Method currently
displays the Ciphering Indicator provided through the SamsungServiceMode methods which unfortunately would
probably be specific to Samsung devices (testing required). Currently investigating the layout of the byte[]
layout to enable further access to these functions and a greater understanding of what the Raw request expects.

Mapviewer updated although an issue still exists when the location table is empty causing a NPE when trying to
overlay stored locations onto the Map, this is currently being worked on to correct this issue.