Service, Boot Completed Receiver, TableView & Femtocatcher Additions
Service has been added to the application and is bound to the main activity allowing for
persistent operation even when the activity itself is not active or has been killed, code
requires some further clean up and additions but currently functions as intended. This addition
resulted in the removal of the Device class I created with most functions now residing within
the service.

To allow the service to run once the device has booted a boot completed receiver has been added,
presently this will start the service without the option to disable this functionality but I will
add a settings preference fragment to allow user selection of this function. (Maybe default to
enabled and allow deactiviation?)

I really didn't like the old Textview display and code that powered the details screen shown by
the application so I replaced this with a TableView which hopefully make the app look a little
better.

Femtocatcher methods has been added and modified slightly (basically just to compile) but will be
futher extended and an alert system incorporated to allow action to be taken if connected to a
Femtocell. This code has been copied and modifed from https://github.com/iSECPartners/femtocatcher
with appropriate code comments as required. Further work is required on this in conjunction with
the RIL Raw work but combined will provide a valuable detection system.

The Femtocatcher methods presently only operate on CDMA devices as I have been unable to locate
information regarding the Unique ID ranges used by GSM femtocell's, any details that can be provided
regarding GSM femtocell unique ID values would greatly assist with extending this detection.