Database Helper Updates, OpenCellID database and Code Cleanup
Major update of the database helper functions to significantly improve the insert and update of records
within all database tables.

*** NOTE: Database updates will result in the loss of ALL existing data export current data if required
and I will shortly implement the ability to import the exported data.

OpenCellID database table added and this will be populated/updated whenever OpenCellID data is requested
and successfully received, this will allow analysis of existing data such as confirmed cell id's and
average signal strength to assist in IMSI detection.

Removed the CMDProcessor as this was not being utilised at all except for a couple of helper methods which
have been moved into the main AIMSICD path, this can added back at a later stage if required as it provides
a very well executed shell management system, also removed some unneeded imports and cleaned up little bits
of code.