Increase Map Zoom Level, Actionbar/Menu tweaks, Database Viewer UI
Increased Map Viewer default zoom level as it was too high, this should now provide a better display of the
current location.

Invalidated the Actionbar/Menu display whenever the main activity is displayed to ensure the correct icons
are displayed for the tracking indication.

Corrected the Signal Strength tracking indicator icon, this one is a good example of why I should not
code whilst drinking bourbon! :)

Tweaked the Database Viewer card display to use a darker colour scheme and gradient on the cards, I think
this looks 100x better the original white crappy display.