Default Location based on Mobile Country Code (MCC)
Map Viewer will now attempt a 3 step process to determine where to initially display the map:
1. Last location held in the location tracking database,
2. Last known location held by the device from either GPS or Network location service
3. **NEW** Based on current MCC the Map Viewer will move to the Capital City based on the MCC

The default locations are provided through the default_mcc_locations.csv file which is found in the
app/src/main/assets folder which lists Country, Mcc, Longitude and Latitude. These details are held within
a database table and populated whenever the database is upgraded. This detail can also be viewed through
the database viewer.