Correct Main Layout, Service Persistence Preference & Map Viewer Preferences
Main layout corrected so all items are aligned correctly give the display a
consistent look.

When I inverted the preference from Kill Persistent Service to Enable Persistent
Service I missed the check in the onDestroy method so this was still acting in the
wrong way and when Enable Persistent Service was selected it would kill the service.
This has now been corrected and operates as intended.

Converted the alert dialog into a Preference Fragment to allow correct visual display
of the currently selected Map Type, other preferences may also be needed in the
future so this would have probably been required down the track.