Drawer UI Style
Massive revamp of the UI for the application I know that this was not a priority
but I actually got quite sick of having to swipe between pages when testing stuff so
decided to change to the newer drawer style UI.

We can customise this further with an image at the top etc and jazz it up a bit more
and it also fits better with the current style of most other newer Android apps.

I hope you all like this style please let me know what you all think.