Neighbouring Cell Fixes & Shutdown Crash
Corrected the list that stores neighbouring cell information so this should now reflect
the correct contents of the neighbouring cells. Cell ID was not being passes correctly
to the Cell class so would not have been displaying a correct value.

Modified the layout for the neighbouring cells so they display now using the Card UI style and allow scrolling of the details.

Slight fix to the onDestory method which should address the NPE some devices had
when the service was destroyed.