AT Command Injection
First thing you need to do before using this is give a BIG THANK YOU to @E3V3A (E:V:A
on xda) for determining how this implementation was possible because without his
perseverance this would not be possible!

ROOT IS REQUIRED!!

Root terminal methods are used to execute the AT Command Injection so this fragment
will check for both ROOT (su binary) and BUSYBOX to confirm both are available on
your device, if not the execution section of the fragment WILL NOT DISPLAY.

If the initial setup works correctly the next step is trying to determine your Ril
Serial Device through the system property ril.libargs if this is successful then
you are in business and the AT Command Injection system is available for you to test.

This will probably NOT WORK on many devices (such as my i9100) but it will provide
details of the failure if available and display them to the screen but also write them
to the file error.txt in the AIMSICD directory of your external storage.

Have fun and try not to break anything.