AIMSICD Shell
Shell class added to provide an ongoing root shell which will allow for the correct
execution of AT Commands and receiving the response, the first implementation would
run the native command in a single instance which is suitable as commands which
redirect serial device output would not function as intended.

AT Command fragment also updated to include the start of an automatic detection system
for possible candidates for AT command serial devices, currently the system will
attempt to draw the system property [rild.libargs] for the serial device. Secondly
ATCI* devices will be checked within /dev/radio as these have been confirmed as
working on MTK devices with the MT6282 chip.