NPE: Possible Workaround for getAllCellInfo
Method getAllCellInfo is encountering an unknown CellInfo type whilst parsing the
results returned by this method, which is a little strange.

Removed the throw and inserted a log to display the error but also provide more details
around what type of CellInfo is being encountered which may help to address this in a
more suitable fashion.