Drawer Updates
Updates to the Drawer UI components to allow removal of all option menu items, the
tracking functions are now available through the first section of the drawer. Items
will dynamically update to reflect the current status of the tracking item and
FemtoCell tracking will only be shown for CDMA devices.

Moved the all the classes & interface files into a dedicated package folder to make
maintenance easier.