Gmaps API for GSM (CellID) to Geolocation
Incorporated the Gmaps undocumented API to translate a GSM CellID into a location value
(longitude & latitude), functions are made through an async task which implements the
AsyncResponse interface allowing an activity to declare itself as the delegate for
processing of the async task response.

Special thanks to @andr3jx for the correct data writing layout and confirmation that
this method works as all examples (including this method) unfortunately result in a
return code of 1 (error) for me which could be due to regional blocking by Google as
reported in some forums or my crappy i9100.

This method has been initially implemented for testing within the OpenCellID update
methods as a fallback when no location can be found from Location Services, this of course can be extended for further uses such as battery saving mode to enable use of
AIMSICD without the need for active GPS.

Primary Scramling Code display has also been added to the device fragment where this
is available.