Database Update
To prepare correctly for contribution to the OpenCellID project the Cell data table
has been updated to include a number of fields that can be provided to OpenCellID
which include the following: Accuracy of reading, speed at time of reading, bearing at time of reading, Mcc, Mnc and string description of network type.

These major changes have required an upgrade of the database version to enable the
new table structure to be rebuild and as the changes are drastic it will NOT BE
POSSIBLE to restore any previous data.