Neighbouring cell updates + OpenCellID Upload Preference
Reverted the change to the getAllCellInfo method for neighbouring cell data as it
is too unstable and provides very inconsistent data, updated the layout of the
neighbouring cell information to display the PSC, RSSI & Network Type as returned
by the getNeighbouringCells method.

Added a new preference for contribution to the OpenCellID project which will allow
users to contribute tracked cell information to OpenCellID, this is currently only
the preference and the sending of data is still being completed.