Location Services (GPS) Handling
The prompting to enable GPS now better handles the users selection and
detects if GPS was actually activated when taken to the settings
fragment.

The activity itself forces a check as I cannot use the method
startActivityForResult from within the service, so whenever the activity
binds to the service or invokes onResume it will check the GPS location
service is running if Cell Tracking is enabled.

This also forces the user to either enable GPS or cancel the dialog box
which will leave the app in a desired state of operation.