Added TinyDB for easy  persistent variables
- Added TinyDB (courtesy of kcochibili) for easy  use of Shared &
persistent variables. We might not need this, but since it is still
using SharedPreferences and considerably simplifies its use, we may find
this more convenient and transparent. So far it is not used anywhere,
but some code (to be committed later) has some Todo's marked where we
should use it. (i.e. in system property "ocid_downloaded" in RequestTask
and CellTracker).

- trivial code shortening in Helpers
- spelling