Failed attempt to replace using setProp with TinyDB
Tried to replace get/setProp calls with using the simple TinyDB
implementation wrapper to SharedPreferences. I couldn't get this to
work, so someone else need to have a look how to get this done properly.

- Change the alert tag from:
"ALERT: New CID -NOT- found in DBe_import: "
to:
"ALERT: Connected to unknown CID not in DBe_import: "