NC detection fix 1
Fixed NC detection logic and added more Logs to be removed or verbose,
once working.

- Found new bug to be reported:
The code in CellTracker: updateNeighbouringCells() is only executed in
MapViewer and not all the time as needed. It should be updated as soon
as a new CID is connected to or every REFRESH_RATE seconds. And outside
of MapViewer as well...