OCID download progress bar
- Tried to fix the progress bar in the RequestTask Downlaod request,
only to discover that for streamed downloads, aka Chunked Transfer
Encoding, the total size cannot be obtained, which mean that a progress
bar cannot be used. Tried to take care of this gracefully by setting
"total" to an arbitrary value., and progress to half of it. However, we
could use a spinner GIF, but barely worth the effort...
- Added some TAGs.