Newly Formatted and coded SmsDetection Class
Tested this on S5 and Sony Xperia and is working great.

Cut out a lot of duplicate code and changed the way the input stream is
parse so we are now checking line by line rather than sending an array
to check sms function.

seperated code if each sms type for easy error detection.

A lot nicer on the eye than previous class and easier to add too