Add PSC to map marker info window w/ i18n if missing
Noticed that if OCID doesn't contain a PSC value for cells, the app sets PSC to
`666` in the database. Most places will display "invalid" but that's not
translateable.

Refactored the validation method to Cell as static and use it to display the
i18n'd string "Unknown" which I feel best conveys when a PSC would be out-of-
bounds.

Future work should translate unknown string for all langs and use it instead of
the `invalid` string literal.