null-guard service before driving menu options
Pull request #798, solving #794, fails to account for fresh installs & newly
booted phones where the service is not instantiated/running by the time the
menu is created for the first time and a NPE occurs.

If the service is null, simply choose "false" since without a service there is
no way tracking is happening.

Added `onPrepareOptionsMenu` implementation that will redraw the checkboxes
depending on the latest truth every time the menu is open. Not sure if this is
unnecessary.